Создал проект с абстракциями
