+++ Byndyusoft.Logging.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30717.126
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32630.192
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Byndyusoft.Logging", "src\Byndyusoft.Logging\Byndyusoft.Logging.csproj", "{321F072C-4E3C-42FB-9872-265B6ECDE195}"
 EndProject
@@ -9,7 +9,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Byndyusoft.Logging.Sample",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Byndyusoft.Logging.OpenTracing", "src\Byndyusoft.Logging.OpenTracing\Byndyusoft.Logging.OpenTracing.csproj", "{B4C3A10D-21F1-4470-9148-3A60E78956AF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.Logging.OpenTelemetry", "src\Byndyusoft.Logging.OpenTelemetry\Byndyusoft.Logging.OpenTelemetry.csproj", "{B892050A-CAA2-4007-8561-416F686B400C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Byndyusoft.Logging.OpenTelemetry", "src\Byndyusoft.Logging.OpenTelemetry\Byndyusoft.Logging.OpenTelemetry.csproj", "{B892050A-CAA2-4007-8561-416F686B400C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Byndyusoft.Logging.OpenTelemetry.Abstractions", "src\Byndyusoft.Logging.OpenTelemetry.Abstractions\Byndyusoft.Logging.OpenTelemetry.Abstractions.csproj", "{207898BF-5391-4CEF-9327-79BFE73BF794}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +35,10 @@ Global
 		{B892050A-CAA2-4007-8561-416F686B400C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B892050A-CAA2-4007-8561-416F686B400C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B892050A-CAA2-4007-8561-416F686B400C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{207898BF-5391-4CEF-9327-79BFE73BF794}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{207898BF-5391-4CEF-9327-79BFE73BF794}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{207898BF-5391-4CEF-9327-79BFE73BF794}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{207898BF-5391-4CEF-9327-79BFE73BF794}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
