+++ src/Byndyusoft.Logging.OpenTelemetry.Abstractions/LogEventPropertyNames.cs
@@ -1,0 +1,7 @@
+﻿namespace Byndyusoft.Logging
+{
+    public static class LogEventPropertyNames
+    {
+        public static string StructureActivityEventName => "_activity_event_name";
+    }
+}
