+++ src/Byndyusoft.Logging.OpenTelemetry.Abstractions/StructuredActivityEventItem.cs
@@ -1,0 +1,18 @@
+﻿namespace Byndyusoft.Logging
+{
+    public class StructuredActivityEventItem
+    {
+        public string Name { get; }
+
+        public object Value { get; }
+
+        public string Description { get; }
+
+        public StructuredActivityEventItem(string name, object value, string description)
+        {
+            Name = name;
+            Value = value;
+            Description = description;
+        }
+    }
+}
